refactor: split llm/tools.ts monolith into a per-tool registry (#982)
The 2157-line tools.ts bundled a 724-line NOTEBASE_TOOLS definition array,
a 21-case executeNotebaseTool switch, and ~40 run/parse functions. Adding
one LLM tool meant three coordinated edits across the monolith — the OCP
friction the skills pipeline next door doesn't have.

Split into src/main/llm/tools/ with one self-registering { definition, run }
module per tool (20 default tools + ask-user), composed by registry.ts:
- executeNotebaseTool is now a DISPATCH lookup, not a switch.
- NOTEBASE_TOOLS derives from the registry (order preserved).
- types.ts holds ToolContext/ToolCallbacks + new ToolResult/NotebaseTool.
- _shared.ts holds the one genuinely cross-tool helper (runProposeRefactor,
  shared by propose_note_rename + propose_note_move); every other helper
  lives with its single-consumer tool.
- src/main/llm/tools.ts is now a 2-line barrel re-exporting the public API,
  so none of the 12 importers change.

Adding a tool is now: create a file, add one registry entry — no central
switch edit. Logic moved verbatim; behavior preserved.

Verified: pnpm lint clean; full tests/main green (195 files / 1961 tests),
incl. the write-guard trust invariant and the TOOL_CALLBACK_KEYS guard.
